+++ download-video-gui.sh
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: MIT
 # ============================================================================
 # Name        : download-video-gui.sh
-# Version     : 2.1.33
+# Version     : 2.1.34
 # Date        : 2026-08-23
 # Description : Zenity GUI for MKV video, authenticated YouTube HLS, or audio.
 # ============================================================================